refactor(mcp): split HTTP transport, migrate pairing store to actor, add SSE keep-alive
