+++ TablePro/Core/MCP/Transport/MCPHttpServerTransport.swift
@@ -21,6 +21,7 @@ public actor MCPHttpServerTransport {
 
     private var listener: NWListener?
     private var connections: [UUID: HttpConnectionContext] = [:]
+    private var sseWriters: [UUID: MCPSseWriter] = [:]
     private var sseConnectionsBySession: [MCPSessionId: UUID] = [:]
     private var sessionEventsTask: Task<Void, Never>?
 
@@ -63,7 +64,7 @@ public actor MCPHttpServerTransport {
         Self.logger.info("Starting MCP HTTP server: bind=\(String(describing: self.configuration.bindAddress)) port=\(self.configuration.port) tls=\(self.configuration.tls != nil)")
 
         if configuration.bindAddress == .anyInterface, configuration.tls == nil {
-            Self.logger.error("Remote access requested without TLS — refusing to start")
+            Self.logger.error("Remote access requested without TLS, refusing to start")
             throw MCPHttpServerError.tlsRequiredForRemoteAccess
         }
 
@@ -100,6 +101,11 @@ public actor MCPHttpServerTransport {
         sessionEventsTask?.cancel()
         sessionEventsTask = nil
 
+        for (_, writer) in sseWriters {
+            await writer.stop()
+        }
+        sseWriters.removeAll()
+
         for (_, context) in connections {
             await context.cancel()
         }
@@ -125,14 +131,14 @@ public actor MCPHttpServerTransport {
 
     public func sendNotification(_ notification: JsonRpcNotification, toSession sessionId: MCPSessionId) async {
         guard let connectionId = sseConnectionsBySession[sessionId],
-              let context = connections[connectionId] else {
+              let writer = sseWriters[connectionId] else {
             return
         }
 
         let message = JsonRpcMessage.notification(notification)
         guard let data = try? JsonRpcCodec.encode(message),
               let text = String(data: data, encoding: .utf8) else { return }
-        await context.writeSseFrame(SseFrame(data: text))
+        await writer.writeFrame(SseFrame(data: text))
     }
 
     public func broadcastNotification(_ notification: JsonRpcNotification) async {
@@ -210,8 +216,7 @@ public actor MCPHttpServerTransport {
     }
 
     private func handleSessionTerminated(_ sessionId: MCPSessionId, reason: MCPSessionTerminationReason) async {
-        guard let connectionId = sseConnectionsBySession.removeValue(forKey: sessionId),
-              let context = connections[connectionId] else {
+        guard let connectionId = sseConnectionsBySession.removeValue(forKey: sessionId) else {
             return
         }
 
@@ -228,8 +233,14 @@ public actor MCPHttpServerTransport {
         case .capacityEvicted:
             comment = "capacity-evicted"
         }
-        await context.writeRaw(Data("\u{003A} \(comment)\n\n".utf8))
-        await context.cancel()
+
+        if let writer = sseWriters.removeValue(forKey: connectionId) {
+            await writer.writeComment(comment)
+            await writer.stop()
+        } else if let context = connections[connectionId] {
+            await context.writeRaw(Data("\u{003A} \(comment)\n\n".utf8))
+            await context.cancel()
+        }
         connections.removeValue(forKey: connectionId)
     }
 
@@ -238,41 +249,61 @@ public actor MCPHttpServerTransport {
         Self.logger.debug("Accepted connection \(connectionId, privacy: .public)")
         let context = HttpConnectionContext(id: connectionId, connection: connection)
         connections[connectionId] = context
+        let router = makeRouter()
         await context.start { [weak self] data in
             guard let self else { return }
-            await self.handleReceivedData(connectionId: connectionId, data: data)
+            await self.handleReceivedData(connectionId: connectionId, data: data, router: router)
         } onClosed: { [weak self] in
             guard let self else { return }
             await self.removeConnection(connectionId: connectionId)
         }
     }
 
+    private func makeRouter() -> MCPHttpRequestRouter {
+        let exchangesContinuation = self.exchangesContinuation
+        let transport = self
+        return MCPHttpRequestRouter(
+            configuration: configuration,
+            sessionStore: sessionStore,
+            authenticator: authenticator,
+            clock: clock,
+            emitInbound: { exchange in
+                exchangesContinuation.yield(exchange)
+            },
+            startSse: { connectionId, sessionId, context in
+                await transport.attachSseWriter(connectionId: connectionId, sessionId: sessionId, context: context)
+            },
+            makeResponderSink: { context in
+                TransportResponderSink(transport: transport, context: context)
+            }
+        )
+    }
+
     private func removeConnection(connectionId: UUID) async {
         connections.removeValue(forKey: connectionId)
+        if let writer = sseWriters.removeValue(forKey: connectionId) {
+            await writer.stop()
+        }
         let pairs = sseConnectionsBySession.filter { $0.value == connectionId }
         for (sessionId, _) in pairs {
             sseConnectionsBySession.removeValue(forKey: sessionId)
         }
     }
 
-    private func handleReceivedData(connectionId: UUID, data: Data) async {
+    private func handleReceivedData(connectionId: UUID, data: Data, router: MCPHttpRequestRouter) async {
         guard let context = connections[connectionId] else { return }
 
         let parseResult: HttpRequestParseResult
         do {
             parseResult = try HttpRequestParser.parse(data)
         } catch HttpRequestParseError.bodyTooLarge {
-            await respondTopLevel(context: context, error: .payloadTooLarge(), requestId: nil)
+            await respondParseFailure(context: context, status: .payloadTooLarge)
             return
         } catch HttpRequestParseError.headerTooLarge {
-            await respondTopLevel(context: context, error: .payloadTooLarge(), requestId: nil)
+            await respondParseFailure(context: context, status: .payloadTooLarge)
             return
         } catch {
-            await respondTopLevel(
-                context: context,
-                error: .invalidRequest(detail: "Malformed HTTP"),
-                requestId: nil
-            )
+            await respondParseFailure(context: context, status: .badRequest, detail: "Malformed HTTP")
             return
         }
 
@@ -281,418 +312,18 @@ public actor MCPHttpServerTransport {
             return
         case .complete(let head, let body, _):
             await context.markRequestComplete()
-            await dispatch(head: head, body: body, context: context)
+            await router.dispatch(head: head, body: body, context: context)
         }
     }
 
-    private func dispatch(head: HttpRequestHead, body: Data, context: HttpConnectionContext) async {
-        let clientAddress: MCPClientAddress = await context.clientAddress()
-        let now = await clock.now()
-
-        await context.setOrigin(head.headers.value(for: "Origin"))
-
-        if head.method == .post, stripQueryString(head.path) == "/v1/integrations/exchange" {
-            await handleIntegrationsExchange(body: body, context: context)
-            return
-        }
-
-        switch head.method {
-        case .options:
-            await context.writeOptions204()
-            await context.cancel()
-            return
-        case .get:
-            await handleGetMcp(head: head, context: context, clientAddress: clientAddress)
-        case .post:
-            await handlePostMcp(head: head, body: body, context: context, clientAddress: clientAddress, now: now)
-        case .delete:
-            await handleDeleteMcp(head: head, context: context, clientAddress: clientAddress)
-        default:
-            await respondTopLevel(
-                context: context,
-                error: MCPProtocolError(
-                    code: JsonRpcErrorCode.methodNotFound,
-                    message: "Method not allowed",
-                    httpStatus: .methodNotAllowed
-                ),
-                requestId: nil
-            )
-        }
-    }
-
-    private func handleIntegrationsExchange(body: Data, context: HttpConnectionContext) async {
-        struct ExchangeBody: Decodable {
-            let code: String
-            let codeVerifier: String
-            enum CodingKeys: String, CodingKey {
-                case code
-                case codeVerifier = "code_verifier"
-            }
-        }
-        struct ExchangeResponse: Encodable {
-            let token: String
-        }
-
-        Self.logger.info("Integrations exchange request received (\(body.count, privacy: .public) bytes)")
-        let ip = Self.ipString(for: await context.clientAddress())
-
-        let parsed: ExchangeBody
-        do {
-            parsed = try JSONDecoder().decode(ExchangeBody.self, from: body)
-        } catch {
-            Self.logger.warning("Integrations exchange decode failed: \(error.localizedDescription, privacy: .public)")
-            MCPAuditLogger.logPairingExchange(outcome: .denied, ip: ip, details: "invalid JSON body")
-            await context.writePlainJsonError(status: .badRequest, message: "Invalid JSON body")
-            await context.cancel()
-            return
-        }
-
-        guard !parsed.code.isEmpty, !parsed.codeVerifier.isEmpty else {
-            Self.logger.warning("Integrations exchange missing code or verifier")
-            MCPAuditLogger.logPairingExchange(
-                outcome: .denied,
-                ip: ip,
-                details: "missing code or code_verifier"
-            )
-            await context.writePlainJsonError(status: .badRequest, message: "Missing code or code_verifier")
-            await context.cancel()
-            return
-        }
-
-        guard parsed.code.utf8.count <= 1_024, parsed.codeVerifier.utf8.count <= 1_024 else {
-            Self.logger.warning("Integrations exchange field exceeds size cap")
-            MCPAuditLogger.logPairingExchange(
-                outcome: .denied,
-                ip: ip,
-                details: "field exceeds 1_024 bytes"
-            )
-            await context.writePlainJsonError(status: .badRequest, message: "Field exceeds size limit")
-            await context.cancel()
-            return
-        }
-
-        let exchange = PairingExchange(code: parsed.code, verifier: parsed.codeVerifier)
-        let outcome: Result<String, Error> = await MainActor.run {
-            do {
-                return .success(try MCPPairingService.shared.exchange(exchange))
-            } catch {
-                return .failure(error)
-            }
-        }
-
-        switch outcome {
-        case .success(let token):
-            Self.logger.info("Integrations exchange succeeded (token len=\(token.count, privacy: .public))")
-            let label = await Self.resolveTokenLabel(for: token)
-            MCPAuditLogger.logPairingExchange(outcome: .success, tokenName: label, ip: ip)
-            let payload = (try? JSONEncoder().encode(ExchangeResponse(token: token))) ?? Data()
-            await context.writePlainJsonResponse(status: .ok, body: payload)
-            await context.cancel()
-        case .failure(let error):
-            let mapped = Self.mapExchangeError(error)
-            Self.logger.warning("Integrations exchange failed: status=\(mapped.status.code, privacy: .public) reason=\(mapped.message, privacy: .public)")
-            MCPAuditLogger.logPairingExchange(
-                outcome: .denied,
-                ip: ip,
-                details: mapped.message
-            )
-            await context.writePlainJsonError(status: mapped.status, message: mapped.message)
-            await context.cancel()
-        }
-    }
-
-    private static func ipString(for address: MCPClientAddress) -> String {
-        switch address {
-        case .loopback:
-            return "127.0.0.1"
-        case .remote(let host):
-            return host
-        }
-    }
-
-    private static func resolveTokenLabel(for plaintext: String) async -> String? {
-        let store: MCPTokenStore? = await MainActor.run { MCPServerManager.shared.tokenStore }
-        guard let store else { return nil }
-        return await store.validate(bearerToken: plaintext)?.name
-    }
-
-    private static func mapExchangeError(_ error: Error) -> (status: HttpStatus, message: String) {
-        guard let domainError = error as? MCPDataLayerError else {
-            return (.internalServerError, "Internal error")
-        }
-        switch domainError {
-        case .notFound:
-            return (.notFound, "Pairing code not found")
-        case .expired:
-            return (HttpStatus(code: 410, reasonPhrase: "Gone"), "Pairing code expired")
-        case .forbidden:
-            return (.forbidden, "Challenge mismatch")
-        default:
-            return (.internalServerError, "Internal error")
-        }
-    }
-
-    private func handleGetMcp(
-        head: HttpRequestHead,
-        context: HttpConnectionContext,
-        clientAddress: MCPClientAddress
-    ) async {
-        guard pathMatchesMcp(head.path) else {
-            await respondTopLevel(
-                context: context,
-                error: MCPProtocolError(
-                    code: JsonRpcErrorCode.methodNotFound,
-                    message: "Method not found",
-                    httpStatus: .notFound
-                ),
-                requestId: nil
-            )
-            return
-        }
-
-        guard let sessionIdRaw = head.headers.value(for: "Mcp-Session-Id") else {
-            await respondTopLevel(context: context, error: .missingSessionId(), requestId: nil)
-            return
-        }
-
-        if head.headers.value(for: "Last-Event-ID") != nil {
-            await respondTopLevel(
-                context: context,
-                error: MCPProtocolError(
-                    code: JsonRpcErrorCode.serverError,
-                    message: "SSE event replay is not supported",
-                    httpStatus: .notImplemented
-                ),
-                requestId: nil
-            )
-            return
-        }
-
-        if let accept = head.headers.value(for: "Accept"),
-           !accept.lowercased().contains("text/event-stream"),
-           !accept.contains("*/*") {
-            await respondTopLevel(context: context, error: .notAcceptable(), requestId: nil)
-            return
-        }
-
-        let authResult = await authenticate(headers: head.headers, clientAddress: clientAddress)
-        guard case .allow = authResult else {
-            if case .deny(let error) = authResult {
-                await respondTopLevel(context: context, error: error, requestId: nil)
-            }
-            return
-        }
-
-        let sessionId = MCPSessionId(sessionIdRaw)
-        guard await sessionStore.session(id: sessionId) != nil else {
-            await respondTopLevel(context: context, error: .sessionNotFound(), requestId: nil)
-            return
-        }
-
-        await sessionStore.touch(id: sessionId)
-
-        registerSseConnection(connectionId: context.id, sessionId: sessionId)
-        await context.writeSseStreamHeaders(sessionId: sessionId)
-        Self.logger.info("Registered SSE notification stream for session \(sessionId.rawValue, privacy: .public)")
-    }
-
-    private func handlePostMcp(
-        head: HttpRequestHead,
-        body: Data,
-        context: HttpConnectionContext,
-        clientAddress: MCPClientAddress,
-        now: Date
-    ) async {
-        guard pathMatchesMcp(head.path) else {
-            await respondTopLevel(
-                context: context,
-                error: MCPProtocolError(
-                    code: JsonRpcErrorCode.methodNotFound,
-                    message: "Method not found",
-                    httpStatus: .notFound
-                ),
-                requestId: nil
-            )
-            return
-        }
-
-        if body.count > configuration.limits.maxRequestBodyBytes {
-            await respondTopLevel(context: context, error: .payloadTooLarge(), requestId: nil)
-            return
-        }
-
-        let authResult = await authenticate(headers: head.headers, clientAddress: clientAddress)
-        guard case .allow(let principal) = authResult else {
-            if case .deny(let error) = authResult {
-                await respondTopLevel(context: context, error: error, requestId: nil)
-            }
-            return
-        }
-
-        let message: JsonRpcMessage
-        do {
-            message = try JsonRpcCodec.decode(body)
-        } catch {
-            await respondTopLevel(
-                context: context,
-                error: .parseError(detail: String(describing: error)),
-                requestId: nil
-            )
-            return
-        }
-
-        let requestId = extractRequestId(from: message)
-        let methodName = extractMethod(from: message)
-        let mcpProtocolVersion = head.headers.value(for: "mcp-protocol-version")
-
-        let sessionId: MCPSessionId?
-        if methodName == "initialize" {
-            do {
-                let session = try await sessionStore.create()
-                sessionId = session.id
-            } catch {
-                await respondTopLevel(
-                    context: context,
-                    error: .serviceUnavailable(),
-                    requestId: requestId
-                )
-                return
-            }
+    private func respondParseFailure(context: HttpConnectionContext, status: HttpStatus, detail: String? = nil) async {
+        let error: MCPProtocolError
+        if status.code == HttpStatus.payloadTooLarge.code {
+            error = .payloadTooLarge()
         } else {
-            guard let raw = head.headers.value(for: "Mcp-Session-Id") else {
-                await respondTopLevel(context: context, error: .missingSessionId(), requestId: requestId)
-                return
-            }
-            let candidate = MCPSessionId(raw)
-            guard let session = await sessionStore.session(id: candidate) else {
-                await respondTopLevel(context: context, error: .sessionNotFound(), requestId: requestId)
-                return
-            }
-            if let mismatch = await Self.protocolVersionMismatch(
-                session: session,
-                headerValue: mcpProtocolVersion
-            ) {
-                await respondTopLevel(context: context, error: mismatch, requestId: requestId)
-                return
-            }
-            sessionId = candidate
-            await sessionStore.touch(id: candidate)
+            error = .invalidRequest(detail: detail ?? "Bad request")
         }
-
-        let sink = TransportResponderSink(transport: self, context: context)
-        let responder = MCPExchangeResponder(sink: sink, requestId: requestId)
-
-        let exchangeContext = MCPInboundContext(
-            sessionId: sessionId,
-            principal: principal,
-            clientAddress: clientAddress,
-            receivedAt: now,
-            mcpProtocolVersion: mcpProtocolVersion
-        )
-        let exchange = MCPInboundExchange(
-            message: message,
-            context: exchangeContext,
-            responder: responder
-        )
-        let yieldResult = exchangesContinuation.yield(exchange)
-        if case .dropped = yieldResult {
-            Self.logger.warning("exchanges buffer full, dropped inbound message — dispatcher is falling behind")
-        }
-    }
-
-    private func handleDeleteMcp(
-        head: HttpRequestHead,
-        context: HttpConnectionContext,
-        clientAddress: MCPClientAddress
-    ) async {
-        guard pathMatchesMcp(head.path) else {
-            await respondTopLevel(
-                context: context,
-                error: MCPProtocolError(
-                    code: JsonRpcErrorCode.methodNotFound,
-                    message: "Method not found",
-                    httpStatus: .notFound
-                ),
-                requestId: nil
-            )
-            return
-        }
-
-        let authResult = await authenticate(headers: head.headers, clientAddress: clientAddress)
-        guard case .allow = authResult else {
-            if case .deny(let error) = authResult {
-                await respondTopLevel(context: context, error: error, requestId: nil)
-            }
-            return
-        }
-
-        guard let raw = head.headers.value(for: "Mcp-Session-Id") else {
-            await respondTopLevel(context: context, error: .missingSessionId(), requestId: nil)
-            return
-        }
-
-        let sessionId = MCPSessionId(raw)
-        guard await sessionStore.session(id: sessionId) != nil else {
-            await respondTopLevel(context: context, error: .sessionNotFound(), requestId: nil)
-            return
-        }
-
-        await sessionStore.terminate(id: sessionId, reason: .clientRequested)
-        await context.writeNoContent()
-        await context.cancel()
-    }
-
-    private func authenticate(
-        headers: HttpHeaders,
-        clientAddress: MCPClientAddress
-    ) async -> AuthResult {
-        let authHeader = headers.value(for: "Authorization")
-        let decision = await authenticator.authenticate(
-            authorizationHeader: authHeader,
-            clientAddress: clientAddress
-        )
-        switch decision {
-        case .allow(let principal):
-            return .allow(principal)
-        case .deny(let reason):
-            let mcpError = mapDenialToProtocolError(reason)
-            return .deny(mcpError)
-        }
-    }
-
-    private func mapDenialToProtocolError(_ reason: MCPAuthDenialReason) -> MCPProtocolError {
-        switch reason.httpStatus {
-        case 401:
-            if let challenge = reason.challenge {
-                if challenge.contains("invalid_token") {
-                    if challenge.contains("token_expired") || challenge.contains("token expired") {
-                        return .tokenExpired()
-                    }
-                    return .tokenInvalid()
-                }
-                return .unauthenticated(challenge: challenge)
-            }
-            return .unauthenticated()
-        case 403:
-            return .forbidden(reason: reason.logMessage)
-        case 429:
-            return .rateLimited(retryAfterSeconds: reason.retryAfterSeconds)
-        default:
-            return MCPProtocolError(
-                code: JsonRpcErrorCode.serverError,
-                message: reason.logMessage,
-                httpStatus: HttpStatus(code: reason.httpStatus, reasonPhrase: "Error"),
-                extraHeaders: reason.challenge.map { [("WWW-Authenticate", $0)] } ?? []
-            )
-        }
-    }
-
-    private func respondTopLevel(
-        context: HttpConnectionContext,
-        error: MCPProtocolError,
-        requestId: JsonRpcId?
-    ) async {
-        let envelope = error.toJsonRpcErrorResponse(id: requestId)
+        let envelope = error.toJsonRpcErrorResponse(id: nil)
         let data = (try? JSONEncoder().encode(envelope)) ?? Data()
         await context.writeJsonResponse(
             data: data,
@@ -703,306 +334,28 @@ public actor MCPHttpServerTransport {
         await context.cancel()
     }
 
-    private func pathMatchesMcp(_ path: String) -> Bool {
-        let trimmed = stripQueryString(path)
-        return trimmed == "/mcp" || trimmed == "/mcp/"
-    }
-
-    private static func protocolVersionMismatch(
-        session: MCPSession,
-        headerValue: String?
-    ) async -> MCPProtocolError? {
-        let state = await session.state
-        guard case .ready = state else { return nil }
-        guard let negotiated = await session.negotiatedProtocolVersion else { return nil }
-        guard let headerValue, !headerValue.isEmpty else { return nil }
-        if headerValue == negotiated { return nil }
-        return .invalidRequest(
-            detail: "MCP-Protocol-Version mismatch: client sent \(headerValue), session negotiated \(negotiated)"
-        )
-    }
-
-    private func stripQueryString(_ path: String) -> String {
-        if let questionIndex = path.firstIndex(of: "?") {
-            return String(path[path.startIndex..<questionIndex])
-        }
-        return path
-    }
-
-    private func extractRequestId(from message: JsonRpcMessage) -> JsonRpcId? {
-        switch message {
-        case .request(let request):
-            return request.id
-        case .successResponse(let response):
-            return response.id
-        case .errorResponse(let response):
-            return response.id
-        case .notification:
-            return nil
-        }
-    }
-
-    private func extractMethod(from message: JsonRpcMessage) -> String? {
-        switch message {
-        case .request(let request):
-            return request.method
-        case .notification(let notification):
-            return notification.method
-        case .successResponse, .errorResponse:
-            return nil
-        }
-    }
-
-    fileprivate func registerSseConnection(connectionId: UUID, sessionId: MCPSessionId) {
-        if let previous = sseConnectionsBySession[sessionId], previous != connectionId,
-           let oldContext = connections[previous] {
-            Task { await oldContext.cancel() }
+    fileprivate func attachSseWriter(
+        connectionId: UUID,
+        sessionId: MCPSessionId,
+        context: HttpConnectionContext
+    ) async {
+        if let previous = sseConnectionsBySession[sessionId], previous != connectionId {
+            if let oldWriter = sseWriters.removeValue(forKey: previous) {
+                await oldWriter.stop()
+            } else if let oldContext = connections[previous] {
+                await oldContext.cancel()
+            }
             connections.removeValue(forKey: previous)
         }
+        let writer = MCPSseWriter(context: context)
+        sseWriters[connectionId] = writer
         sseConnectionsBySession[sessionId] = connectionId
+        await writer.startStream(sessionId: sessionId)
     }
 
-    private enum AuthResult {
-        case allow(MCPPrincipal)
-        case deny(MCPProtocolError)
-    }
-}
-
-actor HttpConnectionContext {
-    private static let logger = Logger(subsystem: "com.TablePro", category: "MCP.HttpServer")
-
-    nonisolated let id: UUID
-    private let connection: NWConnection
-    private var receiveBuffer = Data()
-    private var requestComplete = false
-    private var cancelled = false
-    private var sseActive = false
-    private var origin: String?
-
-    init(id: UUID, connection: NWConnection) {
-        self.id = id
-        self.connection = connection
-    }
-
-    func setOrigin(_ value: String?) {
-        origin = value
-    }
-
-    private func corsHeaders() -> [(String, String)] {
-        MCPCorsHeaders.headers(forOrigin: origin)
-    }
-
-    func start(
-        onData: @escaping @Sendable (Data) async -> Void,
-        onClosed: @escaping @Sendable () async -> Void
-    ) {
-        let nwConnection = connection
-        nwConnection.stateUpdateHandler = { [weak self] state in
-            guard let self else { return }
-            switch state {
-            case .ready:
-                Task { await self.beginReading(onData: onData, onClosed: onClosed) }
-            case .failed:
-                Task { await self.handleClosed(onClosed: onClosed) }
-            case .cancelled:
-                Task { await self.handleClosed(onClosed: onClosed) }
-            default:
-                break
-            }
-        }
-        nwConnection.start(queue: .global(qos: .userInitiated))
-    }
-
-    private func beginReading(
-        onData: @escaping @Sendable (Data) async -> Void,
-        onClosed: @escaping @Sendable () async -> Void
-    ) {
-        scheduleReceive(onData: onData, onClosed: onClosed)
-    }
-
-    private func scheduleReceive(
-        onData: @escaping @Sendable (Data) async -> Void,
-        onClosed: @escaping @Sendable () async -> Void
-    ) {
-        if cancelled || requestComplete { return }
-        connection.receive(minimumIncompleteLength: 1, maximumLength: 65_536) { [weak self] content, _, isComplete, error in
-            guard let self else { return }
-            Task {
-                await self.handleReceive(
-                    content: content,
-                    isComplete: isComplete,
-                    error: error,
-                    onData: onData,
-                    onClosed: onClosed
-                )
-            }
-        }
-    }
-
-    private func handleReceive(
-        content: Data?,
-        isComplete: Bool,
-        error: NWError?,
-        onData: @escaping @Sendable (Data) async -> Void,
-        onClosed: @escaping @Sendable () async -> Void
-    ) async {
-        if let error {
-            Self.logger.debug("Receive error: \(error.localizedDescription, privacy: .public)")
-            cancel()
-            await onClosed()
-            return
-        }
-
-        if let content {
-            receiveBuffer.append(content)
-            await onData(receiveBuffer)
-        }
-
-        if isComplete {
-            cancel()
-            await onClosed()
-            return
-        }
-
-        if !requestComplete, !cancelled {
-            scheduleReceive(onData: onData, onClosed: onClosed)
-        }
-    }
-
-    private func handleClosed(onClosed: @escaping @Sendable () async -> Void) async {
-        if !cancelled {
-            cancelled = true
-        }
-        await onClosed()
-    }
-
-    func markRequestComplete() {
-        requestComplete = true
-    }
-
-    func clientAddress() -> MCPClientAddress {
-        guard let endpoint = connection.currentPath?.remoteEndpoint,
-              case .hostPort(let host, _) = endpoint else {
-            return .loopback
-        }
-        let hostString = "\(host)"
-        if hostString == "127.0.0.1" || hostString == "::1" || hostString.lowercased() == "localhost" {
-            return .loopback
-        }
-        return .remote(hostString)
-    }
-
-    func writeJsonResponse(
-        data: Data,
-        status: HttpStatus,
-        sessionId: MCPSessionId?,
-        extraHeaders: [(String, String)]
-    ) async {
-        if cancelled { return }
-        var headers: [(String, String)] = [
-            ("Content-Type", "application/json"),
-            ("Connection", "close")
-        ]
-        if let sessionId {
-            headers.append(("Mcp-Session-Id", sessionId.rawValue))
-        }
-        headers.append(contentsOf: extraHeaders)
-        headers.append(contentsOf: self.corsHeaders())
-        let head = HttpResponseHead(status: status, headers: HttpHeaders(headers))
-        let payload = HttpResponseEncoder.encode(head, body: data)
-        await send(payload)
-    }
-
-    func writePlainJsonResponse(status: HttpStatus, body: Data) async {
-        if cancelled { return }
-        var headers: [(String, String)] = [
-            ("Content-Type", "application/json"),
-            ("Connection", "close")
-        ]
-        headers.append(contentsOf: self.corsHeaders())
-        let head = HttpResponseHead(status: status, headers: HttpHeaders(headers))
-        let payload = HttpResponseEncoder.encode(head, body: body)
-        await send(payload)
-    }
-
-    func writePlainJsonError(status: HttpStatus, message: String) async {
-        struct ErrorBody: Encodable { let error: String }
-        let payload = (try? JSONEncoder().encode(ErrorBody(error: message))) ?? Data()
-        await writePlainJsonResponse(status: status, body: payload)
-    }
-
-    func writeOptions204() async {
-        if cancelled { return }
-        var headers: [(String, String)] = [("Connection", "close")]
-        headers.append(contentsOf: self.corsHeaders())
-        let head = HttpResponseHead(status: .noContent, headers: HttpHeaders(headers))
-        let payload = HttpResponseEncoder.encode(head, body: nil)
-        await send(payload)
-    }
-
-    func writeNoContent() async {
-        if cancelled { return }
-        var headers: [(String, String)] = [("Connection", "close")]
-        headers.append(contentsOf: self.corsHeaders())
-        let head = HttpResponseHead(status: .noContent, headers: HttpHeaders(headers))
-        let payload = HttpResponseEncoder.encode(head, body: nil)
-        await send(payload)
-    }
-
-    func writeAccepted() async {
-        if cancelled { return }
-        var headers: [(String, String)] = [("Connection", "close")]
-        headers.append(contentsOf: self.corsHeaders())
-        let head = HttpResponseHead(status: .accepted, headers: HttpHeaders(headers))
-        let payload = HttpResponseEncoder.encode(head, body: nil)
-        await send(payload)
-    }
-
-    func writeSseStreamHeaders(sessionId: MCPSessionId) async {
-        if cancelled { return }
-        sseActive = true
-        var headers: [(String, String)] = [
-            ("Content-Type", "text/event-stream"),
-            ("Cache-Control", "no-cache"),
-            ("Connection", "keep-alive"),
-            ("Mcp-Session-Id", sessionId.rawValue)
-        ]
-        headers.append(contentsOf: self.corsHeaders())
-        let head = HttpResponseHead(status: .ok, headers: HttpHeaders(headers))
-        let payload = HttpResponseEncoder.encode(head, body: nil)
-        await send(payload)
-    }
-
-    func writeSseFrame(_ frame: SseFrame) async {
-        if cancelled { return }
-        let data = SseEncoder.encode(frame)
-        await send(data)
-    }
-
-    func writeRaw(_ data: Data) async {
-        if cancelled { return }
-        await send(data)
-    }
-
-    func cancel() {
-        if cancelled { return }
-        cancelled = true
-        connection.cancel()
-    }
-
-    func isSseActive() -> Bool {
-        sseActive
-    }
-
-    private func send(_ data: Data) async {
-        await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-            connection.send(content: data, completion: .contentProcessed { error in
-                if let error {
-                    Self.logger.debug("Send error: \(error.localizedDescription, privacy: .public)")
-                }
-                continuation.resume()
-            })
-        }
+    fileprivate func registerSseConnection(connectionId: UUID, sessionId: MCPSessionId) async {
+        guard let context = connections[connectionId] else { return }
+        await attachSseWriter(connectionId: connectionId, sessionId: sessionId, context: context)
     }
 }
